Rename Survey.js to SurveyListPage.js
